+++ test/dummy/app/controllers/dashboards_controller.rb
@@ -2,5 +2,6 @@ class DashboardsController < ApplicationController
   def show
     @snapshot = Janela::Snapshot.order(:taken_at).last
     @frame = Janela::Frame.order(:id).first
+    @fixed = params[:status] ? { status_eq: params[:status] } : {}
   end
 end