+++ test/dummy/app/controllers/frames_controller.rb
@@ -1,5 +1,6 @@
 class FramesController < ApplicationController
   def show
     @frame = policy_scope(Janela::Frame).find(params[:id])
+    @fixed = params[:status] ? { status_eq: params[:status] } : {}
   end
 end